refactor: LottoOptions Util 객체 추출
- LottoGenerator에서 사용되고있는 옵션값들 클래스로 추출해서 관리
- LottoGenerator -> LottoNumberAutoGenerator로 변경
