+++ src/main/java/lotto/utils/LottoOptions.java
@@ -1,0 +1,7 @@
+package lotto.utils;
+
+public class LottoOptions {
+    public static final int MIN_NUMBER = 1;
+    public static final int MAX_NUMBER = 45;
+    public static final int LOTTO_SIZE = 6;
+}
